Move RedactedValuesLogger to Calamari.Common and make it the default and registered logger.
