+++ source/Calamari.Common/CalamariFlavourProgram.cs
@@ -25,13 +25,13 @@ using Calamari.Common.Plumbing.Variables;
 
 namespace Calamari.Common
 {
-    public abstract class CalamariFlavourProgram
+    public abstract class  CalamariFlavourProgram
     {
         protected readonly ILog Log;
 
         protected CalamariFlavourProgram(ILog log)
         {
-            Log = log;
+            Log = new RedactedValuesLogger(log);
         }
 
         protected virtual int Run(string[] args)
@@ -39,7 +39,7 @@ namespace Calamari.Common
             try
             {
                 AppDomainConfiguration.SetDefaultRegexMatchTimeout();
-                
+
                 SecurityProtocols.EnableAllSecurityProtocols();
                 var options = CommonOptions.Parse(args);
